fix(desktop): split AccountPane provider-view builder into presentation module
AccountPane.tsx grew to 453 lines after adding the invitee-accept UI,
tripping the frontend-structure LARGE_FRONTEND_FILE check (>400 lines).
Move the pure buildAccountProviderViews helper into
account-profile-presentation.ts alongside its sibling presentation
helpers, with unit test coverage, bringing the component back under
the documented threshold.
